+++ CHANGELOG.md
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `MutableFlags` and `DomainID` fields to `MPTokenIssuance` ledger entry type with ledger-state mutable flags constants (`Lsmf` prefix) and flag setter methods.
 - Added `MutableFlags` helper function in `types` package.
 - Added integration tests for account transactions `AccountSet` and `AccountDelete`
-- Added integration tests for did transaction `DIDSet`
+- Added integration tests for did transactions `DIDSet` and `DIDDelete`
 
 ### Changed